third_party/WebKit/LayoutTests/hittesting/culled-inline-expected.txt

Hello World
To manually test, hover between the two words to see if background color turns green.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS element.id is "culled"
PASS successfullyParsed is true

TEST COMPLETE
